stage4-model-integration-with-mcp-and-llama-stack/mcp-servers/database-mcp/requirements.txt
flask==3.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0
